Web-Development/MyProject/requirements.txt
django
djangorestframework